requirements.txt
chainlit==1.3.2
fpdf==1.7.2
importlib-resources==6.4.0
jaraco.collections==5.1.0
langchain-community==0.3.11
langchain-openai==0.2.12
langgraph==0.2.59
pandas==2.2.3
platformdirs==4.2.2
src==0.0.1
